backend/requirements/requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.2
django-allauth==0.35.0
django-cors-headers==2.1.0
django-rest-auth==0.9.3
djangorestframework==3.7.7
djangorestframework-csv==2.1.0
djangorestframework-jwt==1.11.0
idna==2.6
kombu==4.1.0
oauthlib==2.0.6
psycopg2==2.7.4
PyJWT==1.5.3
python-telegram-handler==2.0.2
python3-openid==3.1.0
pytz==2018.3
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
unicodecsv==0.14.1
urllib3==1.22
vine==1.1.4
xlrd==1.1.0
